% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hence_plots.R
\name{hence_histogram}
\alias{hence_histogram}
\title{Plot a histogram}
\usage{
hence_histogram(
  p_dados,
  p_fill = "#300030",
  p_xlab = "",
  p_ylab = "",
  p_title = "",
  p_subtitle = ""
)
}
\arguments{
\item{p_dados}{a numeric vector}

\item{p_fill}{hex code to fill histogram bars}

\item{p_xlab}{x axis label}

\item{p_ylab}{y axis label}

\item{p_title}{title}

\item{p_subtitle}{subtitle}
}
\description{
Plot a histogram
}
