% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hence_plots.R
\name{hence_line_chart}
\alias{hence_line_chart}
\title{Plot line charts}
\usage{
hence_line_chart(
  p_dados,
  p_x,
  p_y,
  p_color = NULL,
  p_group = NULL,
  p_color_value = "#300030",
  p_colors_values = c("#300030", "#e28743"),
  p_xlab = "",
  p_ylab = "",
  p_title = "",
  p_subtitle = ""
)
}
\arguments{
\item{p_dados}{a data.table}

\item{p_x}{varible to be displayed in x axis}

\item{p_y}{variable to be displayed in y axis}

\item{p_color}{variable to color lines}

\item{p_group}{variable to group lines}

\item{p_xlab}{x axis label}

\item{p_ylab}{y axis label}

\item{p_title}{title of the chart}

\item{p_subtitle}{subtitle of the chart}

\item{p_fill}{hex to fill line chart}
}
\description{
Plot line charts
}
