% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hence_plots.R
\name{hence_table}
\alias{hence_table}
\title{Build a beautiful looking table}
\usage{
hence_table(p_dados, p_align = "c", p_font_size = 18, p_row_names = FALSE)
}
\arguments{
\item{p_dados}{a data.frame or data.table}

\item{p_align}{alignment of text inside table ("c" for center,"r" for right and "l" for left)}

\item{p_row_names}{logical indicating wether or not to include row names in table}
}
\description{
Build a beautiful looking table
}
