% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_null}
\alias{safe_null}
\title{check whether an object is null, but returns FALSE if error occurs}
\usage{
safe_null(x)
}
\description{
check whether an object is null, but returns FALSE if error occurs
}
