% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hence.R
\name{theme_hence}
\alias{theme_hence}
\title{ggplot theme using Hence standards}
\usage{
theme_hence(p_base_size = 15, p_pase_family = "Arial", p_line_colour = "grey")
}
\arguments{
\item{p_line_colour}{}
}
\description{
ggplot theme using Hence standards
}
